I have added the actions to be printed to the console.This includes things such as the weight in decimal format,converted to a double and a character datatype.

+++ src/ProjectClass.java
@@ -4,7 +4,7 @@ public class ProjectClass {
         float myWeightEarthNhlanhla = 60f;
         /*This is the calculation of my weight on Mars by taking 38% of my Earth weight */
         float myWeightMarsNhlanhla  = ( myWeightEarthNhlanhla * 38)/100;
-        double myWeightMarsInDouble = (double)myWeightMarsNhlanhla;
+        double myWeightMarsInDouble = myWeightMarsNhlanhla;
         int myWeightMarsInInt = (int)myWeightMarsNhlanhla;
         char myWeightMarsInChar = (char)myWeightMarsInInt;
         char calculation = (char) (myWeightMarsInChar + '5'+'a');
